requirements.txt
gitdb==4.0.10
GitPython==3.1.31
PyYAML==6.0
smmap==5.0.0
pandas==2.0.1
matplotlib==3.7.1